requirements.txt
tensorflow
scikit-learn
seaborn
matplotlib
numpy
streamlit
Pillow